Minimal Implementation of EAP
