--- docstring_task/abs_value_num_passes_docstring.json
@@ -1,1 +1,0 @@
-{"0.005": 936, "0.01": 862, "0.015": 804, "0.02": 697, "0.025": 617, "0.030000000000000002": 573, "0.034999999999999996": 513, "0.04": 425, "0.045": 381, "0.049999999999999996": 381, "0.055": 378, "0.06": 356, "0.065": 359, "0.07": 355, "0.07500000000000001": 345, "0.08": 335, "0.085": 312, "0.09000000000000001": 299, "0.095": 298, "0.1": 289, "0.10500000000000001": 277, "0.11": 228, "0.115": 223, "0.12000000000000001": 218, "0.125": 218, "0.13": 217, "0.135": 217, "0.14": 198, "0.14500000000000002": 184, "0.15": 160, "0.155": 160}
