--- greaterthan_task/results/greaterthan_absval_num_passes.json
@@ -1,1 +1,0 @@
-{"0.007925": 288, "0.009176315789473683": 285, "0.010427631578947369": 262, "0.011678947368421053": 253, "0.012930263157894736": 238, "0.01418157894736842": 207, "0.015432894736842104": 201, "0.016684210526315787": 201, "0.01793552631578947": 201, "0.019186842105263155": 201, "0.02043815789473684": 151, "0.021689473684210526": 151, "0.02294078947368421": 144, "0.024192105263157897": 144, "0.02544342105263158": 121, "0.026694736842105264": 121, "0.027946052631578948": 121, "0.02919736842105263": 121, "0.030448684210526315": 121, "0.0317": 121}
