--- greaterthan_task/results/greaterthan_absval_pruned_heads.json
@@ -1,1 +1,0 @@
-{"0.007925": [["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<a6.1>", "<a10.4>", "<resid_post>", "<a9.6>", "<m11>", "<m8>", "<m1>", "<m0>", "<a0.5>", "embed", "<m7>", "<m10>", "<a0.1>", "<m5>", "<a6.9>", "<a9.1>", "<m4>", "<m9>"], ["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<a6.1>", "<resid_post>", "<a9.6>", "<m11>", "<m8>", "<m1>", "<m0>", "<a0.5>", "embed", "<m7>", "<m10>", "<a0.1>", "<a6.9>", "<a9.1>", "<m4>", "<m9>"]], "0.009176315789473683": [["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<a6.1>", "<resid_post>", "<a9.6>", "<m11>", "<m8>", "<m1>", "<m0>", "<a0.5>", "embed", "<m7>", "<m10>", "<a0.1>", "<m5>", "<a6.9>", "<a9.1>", "<m4>", "<m9>"], ["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<a6.1>", "<resid_post>", "<a9.6>", "<m11>", "<m8>", "<m1>", "<m0>", "<a0.5>", "embed", "<m7>", "<m10>", "<a0.1>", "<a6.9>", "<a9.1>", "<m4>", "<m9>"]], "0.010427631578947369": [["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<a6.1>", "<resid_post>", "<a9.6>", "<m11>", "<m8>", "<m1>", "<m0>", "<a0.5>", "embed", "<m7>", "<m10>", "<a0.1>", "<m5>", "<a6.9>", "<a9.1>", "<m4>", "<m9>"], ["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<a6.1>", "<resid_post>", "<a9.6>", "<m11>", "<m8>", "<m1>", "<m0>", "<a0.5>", "embed", "<m10>", "<a0.1>", "<a6.9>", "<a9.1>", "<m9>"]], "0.011678947368421053": [["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<a6.1>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<a0.1>", "<m5>", "<a6.9>", "<a9.1>", "<m4>", "<m9>"], ["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<a6.1>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<a0.1>", "<a6.9>", "<a9.1>", "<m4>", "<m9>"]], "0.012930263157894736": [["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<a6.1>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<a0.1>", "<m5>", "<a9.1>", "<m4>", "<m9>"], ["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<a6.1>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<a0.1>", "<m5>", "<a9.1>", "<m4>", "<m9>"]], "0.01418157894736842": [["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<a0.1>", "<m5>", "<a9.1>", "<m4>", "<m9>"], ["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<a0.1>", "<a9.1>", "<m4>", "<m9>"]], "0.015432894736842104": [["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<a0.1>", "<m5>", "<a9.1>", "<m4>", "<m9>"], ["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<a0.1>", "<a9.1>", "<m9>"]], "0.016684210526315787": [["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<a0.1>", "<m5>", "<a9.1>", "<m4>", "<m9>"], ["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<a0.1>", "<a9.1>", "<m9>"]], "0.01793552631578947": [["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<a0.1>", "<m5>", "<a9.1>", "<m4>", "<m9>"], ["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<a0.1>", "<a9.1>", "<m9>"]], "0.019186842105263155": [["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<a0.1>", "<m5>", "<a9.1>", "<m4>", "<m9>"], ["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<a0.1>", "<a9.1>", "<m9>"]], "0.02043815789473684": [["<a5.5>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<m5>", "<a9.1>", "<m4>", "<m9>"], ["embed", "<a5.5>", "<resid_post>", "<m6>", "<m11>", "<m8>", "<a7.10>", "<m1>", "<m0>", "<m3>", "<m2>", "<m10>", "<a9.1>", "<m4>", "<a8.8>", "<m9>"]], "0.021689473684210526": [["<a5.5>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<m5>", "<a9.1>", "<m4>", "<m9>"], ["embed", "<a5.5>", "<resid_post>", "<m6>", "<m11>", "<m8>", "<a7.10>", "<m1>", "<m0>", "<m3>", "<m2>", "<m10>", "<a9.1>", "<m4>", "<a8.8>", "<m9>"]], "0.02294078947368421": [["embed", "<a5.5>", "<resid_post>", "<m7>", "<m5>", "<m6>", "<m11>", "<a9.1>", "<m8>", "<a7.10>", "<m1>", "<m3>", "<a10.7>", "<m10>", "<m2>", "<m9>", "<m0>", "<m4>"], ["embed", "<a5.5>", "<resid_post>", "<m7>", "<m6>", "<m11>", "<m8>", "<a7.10>", "<m1>", "<m0>", "<m3>", "<a10.7>", "<m10>", "<m2>", "<a9.1>", "<m4>", "<m9>"]], "0.024192105263157897": [["embed", "<a5.5>", "<resid_post>", "<m7>", "<m5>", "<m6>", "<m11>", "<a9.1>", "<m8>", "<a7.10>", "<m1>", "<m3>", "<a10.7>", "<m10>", "<m2>", "<m9>", "<m0>", "<m4>"], ["embed", "<a5.5>", "<resid_post>", "<m7>", "<m6>", "<m11>", "<m8>", "<a7.10>", "<m1>", "<m0>", "<m3>", "<a10.7>", "<m10>", "<m2>", "<a9.1>", "<m4>", "<m9>"]], "0.02544342105263158": [["embed", "<resid_post>", "<m7>", "<m5>", "<m6>", "<m11>", "<a9.1>", "<m8>", "<a7.10>", "<m1>", "<m3>", "<a10.7>", "<m10>", "<m2>", "<m9>", "<m0>", "<m4>"], ["embed", "<resid_post>", "<m7>", "<m6>", "<m11>", "<m8>", "<a7.10>", "<m1>", "<m0>", "<m3>", "<a10.7>", "<m10>", "<m2>", "<a9.1>", "<m9>"]], "0.026694736842105264": [["embed", "<resid_post>", "<m7>", "<m5>", "<m6>", "<m11>", "<a9.1>", "<m8>", "<a7.10>", "<m1>", "<m3>", "<a10.7>", "<m10>", "<m2>", "<m9>", "<m0>", "<m4>"], ["embed", "<resid_post>", "<m7>", "<m6>", "<m11>", "<m8>", "<a7.10>", "<m1>", "<m0>", "<m3>", "<a10.7>", "<m10>", "<m2>", "<a9.1>", "<m9>"]], "0.027946052631578948": [["embed", "<resid_post>", "<m7>", "<m5>", "<m6>", "<m11>", "<a9.1>", "<m8>", "<a7.10>", "<m1>", "<m3>", "<a10.7>", "<m10>", "<m2>", "<m9>", "<m0>", "<m4>"], ["embed", "<resid_post>", "<m7>", "<m6>", "<m11>", "<m8>", "<a7.10>", "<m1>", "<m0>", "<m3>", "<a10.7>", "<m10>", "<m2>", "<a9.1>", "<m9>"]], "0.02919736842105263": [["embed", "<resid_post>", "<m7>", "<m5>", "<m6>", "<m11>", "<a9.1>", "<m8>", "<a7.10>", "<m1>", "<m3>", "<a10.7>", "<m10>", "<m2>", "<m9>", "<m0>", "<m4>"], ["embed", "<resid_post>", "<m7>", "<m6>", "<m11>", "<m8>", "<a7.10>", "<m1>", "<m0>", "<m3>", "<a10.7>", "<m10>", "<m2>", "<a9.1>", "<m9>"]], "0.030448684210526315": [["embed", "<resid_post>", "<m7>", "<m5>", "<m6>", "<m11>", "<a9.1>", "<m8>", "<a7.10>", "<m1>", "<m3>", "<a10.7>", "<m10>", "<m2>", "<m9>", "<m0>", "<m4>"], ["embed", "<resid_post>", "<m7>", "<m6>", "<m11>", "<m8>", "<a7.10>", "<m1>", "<m0>", "<m3>", "<a10.7>", "<m10>", "<m2>", "<a9.1>", "<m9>"]], "0.0317": [["embed", "<resid_post>", "<m7>", "<m5>", "<m6>", "<m11>", "<a9.1>", "<m8>", "<a7.10>", "<m1>", "<m3>", "<a10.7>", "<m10>", "<m2>", "<m9>", "<m0>", "<m4>"], ["embed", "<resid_post>", "<m7>", "<m6>", "<m11>", "<m8>", "<a7.10>", "<m1>", "<m0>", "<m3>", "<a10.7>", "<m10>", "<m2>", "<a9.1>", "<m9>"]]}
