--- greaterthan_task/results/greaterthan_first_pass_num_passes.json
@@ -1,1 +1,0 @@
-{"0.01585": 201}
