--- greaterthan_task/results/greaterthan_first_pass_pruned_attrs.json
@@ -1,1 +1,0 @@
-{"0.01585": {"L0H0": 6.682943785563111e-05, "L0H2": 0.004924206528812647, "L0H3": 0.007153269834816456, "L0H4": 0.0038876731414347887, "L0H5": 0.010861625894904137, "L0H6": 0.00043244220432825387, "L0H7": 0.006177806295454502, "L0H8": 0.0005415635532699525, "L0H9": 0.001762608764693141, "L0H10": 0.003871350083500147, "L0H11": 0.001359875313937664, "L1H0": 0.0009917484130710363, "L1H1": 0.0009105786448344588, "L1H2": 0.00014150900824461132, "L1H3": 0.0020992462523281574, "L1H4": 0.00017145770834758878, "L1H5": 0.007083612494170666, "L1H6": 0.00016919116023927927, "L1H7": 0.0019029075047001243, "L1H8": 4.725949838757515e-05, "L1H9": 4.230861668474972e-05, "L1H10": 0.00037578342016786337, "L1H11": 0.0014855725457891822, "L2H0": 0.0019191649043932557, "L2H1": 0.00214808969758451, "L2H2": 0.0024934974499046803, "L2H3": 0.000635548320133239, "L2H4": 0.004467578139156103, "L2H5": 0.002218085341155529, "L2H6": 0.0012357577215880156, "L2H7": 0.00027195035363547504, "L2H8": 0.00040346378227695823, "L2H9": 0.0014996943064033985, "L2H10": 0.001109811128117144, "L2H11": 0.0006678666104562581, "L3H0": 0.0016048499383032322, "L3H1": 4.179199459031224e-05, "L3H2": 0.000517719890922308, "L3H3": 9.97509341686964e-05, "L3H4": 0.0016818811418488622, "L3H5": 4.4079264625906944e-05, "L3H6": 0.0007969538564793766, "L3H7": 0.0002894916106015444, "L3H8": 0.001279802294448018, "L3H9": 0.001918319845572114, "L3H10": 0.00046112603740766644, "L3H11": 0.001037092413753271, "L4H0": 0.00062256318051368, "L4H1": 0.0007489825366064906, "L4H2": 0.00022702667047269642, "L4H3": 0.00249772472307086, "L4H4": 0.004218580201268196, "L4H5": 0.0035213467199355364, "L4H6": 0.003603001357987523, "L4H7": 0.0022356221452355385, "L4H8": 0.0070498730055987835, "L4H9": 0.0008734700386412442, "L4H10": 0.00341726653277874, "L4H11": 0.0024168887175619602, "L5H0": 0.00030740260262973607, "L5H1": 0.0011553500080481172, "L5H2": 0.0015157873276621103, "L5H3": 0.00036283250665292144, "L5H4": 0.0009005400934256613, "L5H6": 0.00047112395986914635, "L5H7": 1.4019955415278673e-05, "L5H8": 0.0008736909367144108, "L5H9": 0.0006366747547872365, "L5H10": 0.0018326842691749334, "L5H11": 0.0006416576798073947, "L6H0": 0.00028642534743994474, "L6H1": 0.013869956135749817, "L6H2": 0.0002665279898792505, "L6H3": 0.001548783970065415, "L6H4": 0.0009254538454115391, "L6H5": 0.0030923117883503437, "L6H6": 0.0023855315521359444, "L6H7": 0.0006218899507075548, "L6H8": 0.0002659866586327553, "L6H9": 0.012234240770339966, "L6H10": 0.0002466221048962325, "L6H11": 0.0007808533846400678, "L7H0": 0.0006520153256133199, "L7H1": 0.001032261410728097, "L7H2": 0.00048802251694723964, "L7H3": 0.0012545776553452015, "L7H4": 0.000838662323076278, "L7H5": 0.0008072698838077486, "L7H6": 0.0066537377424538136, "L7H7": 0.006869889795780182, "L7H8": 0.0019015122670680285, "L7H9": 0.0009603225043974817, "L7H11": 0.004181566182523966, "L8H0": 0.0007331593660637736, "L8H1": 0.004181223921477795, "L8H2": 0.000742992851883173, "L8H3": 0.0003785094013437629, "L8H4": 0.000790676916949451, "L8H5": 0.0012235743924975395, "L8H6": 0.0020309530664235353, "L8H7": 0.0008130870992317796, "L8H9": 0.0024738118518143892, "L8H10": 0.006889326497912407, "L9H0": 2.074235089821741e-05, "L9H2": 0.0001314622932113707, "L9H3": 0.002187237609177828, "L9H4": 0.0005771390860900283, "L9H5": 0.0027549793012440205, "L9H6": 0.0111205680295825, "L9H7": 0.00046370632480829954, "L9H8": 0.0003021408338099718, "L9H9": 0.005357271991670132, "L9H10": 0.0001644733129069209, "L9H11": 0.0002228685189038515, "L10H0": 0.0011773486621677876, "L10H1": 0.001437217928469181, "L10H2": 0.006685856729745865, "L10H3": 0.002048744587227702, "L10H4": 0.008055365644395351, "L10H5": 7.358308357652277e-05, "L10H6": 0.001933764317072928, "L10H8": 0.000896780751645565, "L10H9": 0.00019626232096925378, "L10H10": 0.0007632666965946555, "L10H11": 2.5391784220119007e-05, "L11H0": 0.0022834488190710545, "L11H1": 0.0003687163698486984, "L11H2": 0.0013249889016151428, "L11H3": 5.9077941841678694e-05, "L11H4": 0.0003791903145611286, "L11H5": 2.9664173780474812e-05, "L11H6": 0.00020392087753862143, "L11H7": 0.00014993123477324843, "L11H8": 0.002310845535248518, "L11H9": 1.3490192941389978e-05, "L11H10": 0.0018624679651111364, "L11H11": 3.710501914611086e-05}}
