--- greaterthan_task/results/greaterthan_first_pass_pruned_heads.json
@@ -1,1 +1,0 @@
-{"0.01585": [["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<a0.1>", "<m5>", "<a9.1>", "<m4>", "<m9>"], ["<a5.5>", "<a8.11>", "<m6>", "<a7.10>", "<m3>", "<a10.7>", "<m2>", "<a8.8>", "<resid_post>", "<m11>", "<m8>", "<m1>", "<m0>", "embed", "<m7>", "<m10>", "<a0.1>", "<a9.1>", "<m9>"]]}
