--- ioi_task/abs_value_num_passes.json
@@ -1,1 +1,0 @@
-{"0.005": 8600, "0.01": 3956, "0.015": 2815, "0.02": 1929, "0.025": 1536, "0.030000000000000002": 1021, "0.034999999999999996": 871, "0.04": 678, "0.045": 617, "0.049999999999999996": 607, "0.055": 410, "0.06": 383, "0.065": 341, "0.07": 322, "0.07500000000000001": 302, "0.08": 282, "0.085": 282, "0.09000000000000001": 282, "0.095": 260, "0.1": 260, "0.10500000000000001": 250, "0.11": 240, "0.115": 234, "0.12000000000000001": 188, "0.125": 188, "0.13": 188, "0.135": 168, "0.14": 161, "0.14500000000000002": 161, "0.15": 151}
