--- ioi_task/noabs_value_num_passes.json
@@ -1,1 +1,0 @@
-{"0.005": 2960, "0.01": 1787, "0.015": 1034, "0.02": 824, "0.025": 691, "0.030000000000000002": 528, "0.034999999999999996": 421, "0.04": 386, "0.045": 386, "0.049999999999999996": 362, "0.055": 331, "0.06": 300, "0.065": 261, "0.07": 231, "0.07500000000000001": 195, "0.08": 199, "0.085": 159, "0.09000000000000001": 154, "0.095": 154, "0.1": 154, "0.10500000000000001": 154, "0.11": 154, "0.115": 154, "0.12000000000000001": 154, "0.125": 154, "0.13": 154, "0.135": 124, "0.14": 113, "0.14500000000000002": 113, "0.15": 113}
